requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.10
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.0
Pillow==11.1.0
python-decouple==3.8
dj-database-url==2.1.0
djangorestframework-simplejwt==5.3.0
django-extensions==3.2.3
sentry-sdk==1.38.0
prometheus-client==0.19.0
django-prometheus==2.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
drf-yasg==1.21.7
